refactor: update authentication imports and restructure auth configuration
