+++ auth.ts
@@ -1,0 +1,2 @@
+import { auth } from "~libs/auth/better-auth-config";
+export default auth;
