+++ src/libs/auth/components/SignOut.tsx
@@ -1,4 +1,4 @@
-import { authReactClient } from "../auth-client";
+import { authReactClient } from "../auth-clients";
 
 export const SignOut = () => {
   return (
