+++ src/pages/api/auth/[...all].ts
@@ -1,5 +1,5 @@
 import type { APIRoute } from "astro";
-import { auth } from "~auth";
+import { auth } from "~libs/auth/better-auth-config";
 
 export const ALL: APIRoute = async (ctx) => {
   return auth.handler(ctx.request);
